Fix build, remove framework detection
