+++ docs/package.json
@@ -10,6 +10,7 @@
     "react-dom": "^16.6.0"
   },
   "scripts": {
-    "watch": "../packages/cli/lib/bin/catalog-start.js ."
+    "watch": "../packages/cli/lib/bin/catalog-start.js .",
+    "build": "../packages/cli/lib/bin/catalog-build.js ."
   }
 }
