+++ packages/cli/package.json
@@ -51,6 +51,7 @@
     "sander": "^0.6.0",
     "sass-loader": "^7.1.0",
     "style-loader": "^0.23.1",
+    "terser-webpack-plugin": "^1.2.2",
     "url-loader": "^1.1.2",
     "webpack": "^4.29.1",
     "webpack-dev-server": "^3.1.14",
