+++ packages/cli/src/actions/loadWebpackConfig.ts
@@ -1,13 +1,14 @@
 import webpack from "webpack";
 import HtmlWebpackPlugin from "html-webpack-plugin";
+import TerserPlugin from "terser-webpack-plugin";
+import MiniCssExtractPlugin from "mini-css-extract-plugin";
 import ManifestPlugin from "webpack-manifest-plugin";
+
+import getCSSModuleLocalIdent from "react-dev-utils/getCSSModuleLocalIdent";
 import WatchMissingNodeModulesPlugin from "react-dev-utils/WatchMissingNodeModulesPlugin";
 import InterpolateHtmlPlugin from "react-dev-utils/InterpolateHtmlPlugin";
 import FriendlyErrorsWebpackPlugin from "friendly-errors-webpack-plugin";
-import { exists } from "sander";
 
-import createReactAppConfig from "../config/createReactApp";
-import nextConfig from "../config/next";
 import getClientEnvironment from "../config/env";
 import { link } from "../utils/format";
 
@@ -20,6 +21,73 @@ type LoadWebpackOptions = {
 };
 type WebpackConfig = {};
 
+const shouldUseSourceMap = true;
+
+// style files regexes
+const cssRegex = /\.css$/;
+const cssModuleRegex = /\.module\.css$/;
+const sassRegex = /\.(scss|sass)$/;
+const sassModuleRegex = /\.module\.(scss|sass)$/;
+
+// FIXME: detect when this should be true
+const shouldUseRelativeAssetPaths = false;
+
+// common function to get style loaders
+const getStyleLoaders = ({
+  cssOptions,
+  preProcessor,
+  dev
+}: {
+  cssOptions: any;
+  preProcessor?: string;
+  dev: boolean;
+}) => {
+  const loaders = [
+    dev && require.resolve("style-loader"),
+    !dev && {
+      loader: MiniCssExtractPlugin.loader,
+      options: Object.assign(
+        {},
+        shouldUseRelativeAssetPaths ? { publicPath: "../../" } : undefined
+      )
+    },
+    {
+      loader: require.resolve("css-loader"),
+      options: cssOptions
+    },
+    {
+      // Options for PostCSS as we reference these options twice
+      // Adds vendor prefixing based on your specified browser support in
+      // package.json
+      loader: require.resolve("postcss-loader"),
+      options: {
+        // Necessary for external CSS imports to work
+        // https://github.com/facebook/create-react-app/issues/2677
+        ident: "postcss",
+        plugins: () => [
+          require("postcss-flexbugs-fixes"),
+          require("postcss-preset-env")({
+            autoprefixer: {
+              flexbox: "no-2009"
+            },
+            stage: 3
+          })
+        ],
+        sourceMap: !dev && shouldUseSourceMap
+      }
+    }
+  ].filter(Boolean);
+  if (preProcessor) {
+    loaders.push({
+      loader: require.resolve(preProcessor),
+      options: {
+        sourceMap: !dev && shouldUseSourceMap
+      }
+    });
+  }
+  return loaders;
+};
+
 export default async ({
   paths,
   framework,
@@ -27,11 +95,6 @@ export default async ({
   url,
   useBabelrc
 }: LoadWebpackOptions): Promise<WebpackConfig> => {
-  const frameworkConfig =
-    framework === "NEXT"
-      ? nextConfig(paths, useBabelrc, dev)
-      : createReactAppConfig(paths, useBabelrc, dev);
-
   const env = getClientEnvironment(paths.publicUrl.replace(/\/$/, ""));
 
   const devPlugins = dev
@@ -93,12 +156,122 @@ export default async ({
     },
     module: {
       rules: [
-        ...frameworkConfig.moduleRules,
         {
-          test: /\.md$/,
-          loaders: [
-            require.resolve("@catalog/markdown-loader"),
-            require.resolve("raw-loader")
+          oneOf: [
+            // "url" loader works like "file" loader except that it embeds assets
+            // smaller than specified limit in bytes as data URLs to avoid requests.
+            // A missing `test` is equivalent to a match.
+            {
+              test: [/\.bmp$/, /\.gif$/, /\.jpe?g$/, /\.png$/],
+              loader: require.resolve("url-loader"),
+              options: {
+                limit: 10000,
+                name: "static/media/[name].[hash:8].[ext]"
+              }
+            },
+            // Process JS with Babel.
+            {
+              test: /\.(mjs|js|jsx|ts|tsx)$/,
+              include: [paths.appRoot, paths.catalogSrcDir],
+              exclude: /(lodash|standalone)/,
+              loader: require.resolve("babel-loader"),
+              options: {
+                babelrc: useBabelrc,
+                presets: useBabelrc
+                  ? []
+                  : [
+                      require.resolve("babel-preset-react-app"),
+                      require.resolve("@catalog/babel-preset")
+                    ],
+                cacheDirectory: true
+              }
+            },
+            // "postcss" loader applies autoprefixer to our CSS.
+            // "css" loader resolves paths in CSS and adds assets as dependencies.
+            // "style" loader turns CSS into JS modules that inject <style> tags.
+            // In production, we use MiniCSSExtractPlugin to extract that CSS
+            // to a file, but in development "style" loader enables hot editing
+            // of CSS.
+            // By default we support CSS Modules with the extension .module.css
+            {
+              test: cssRegex,
+              exclude: cssModuleRegex,
+              use: getStyleLoaders({
+                cssOptions: {
+                  importLoaders: 1,
+                  sourceMap: !dev && shouldUseSourceMap
+                },
+                dev
+              }),
+              // Don't consider CSS imports dead code even if the
+              // containing package claims to have no side effects.
+              // Remove this when webpack adds a warning or an error for this.
+              // See https://github.com/webpack/webpack/issues/6571
+              sideEffects: true
+            },
+            // Adds support for CSS Modules (https://github.com/css-modules/css-modules)
+            // using the extension .module.css
+            {
+              test: cssModuleRegex,
+              use: getStyleLoaders({
+                cssOptions: {
+                  importLoaders: 1,
+                  sourceMap: !dev && shouldUseSourceMap,
+                  modules: true,
+                  getLocalIdent: getCSSModuleLocalIdent
+                },
+                dev
+              })
+            },
+            // Opt-in support for SASS (using .scss or .sass extensions).
+            // By default we support SASS Modules with the
+            // extensions .module.scss or .module.sass
+            {
+              test: sassRegex,
+              exclude: sassModuleRegex,
+              use: getStyleLoaders({
+                cssOptions: {
+                  importLoaders: 2,
+                  sourceMap: !dev && shouldUseSourceMap
+                },
+                preProcessor: "sass-loader",
+                dev
+              }),
+              // Don't consider CSS imports dead code even if the
+              // containing package claims to have no side effects.
+              // Remove this when webpack adds a warning or an error for this.
+              // See https://github.com/webpack/webpack/issues/6571
+              sideEffects: true
+            },
+            // Adds support for CSS Modules, but using SASS
+            // using the extension .module.scss or .module.sass
+            {
+              test: sassModuleRegex,
+              use: getStyleLoaders({
+                cssOptions: {
+                  importLoaders: 2,
+                  sourceMap: !dev && shouldUseSourceMap,
+                  modules: true,
+                  getLocalIdent: getCSSModuleLocalIdent
+                },
+                preProcessor: "sass-loader",
+                dev
+              })
+            },
+            {
+              test: /\.md$/,
+              loaders: [
+                require.resolve("@catalog/markdown-loader"),
+                require.resolve("raw-loader")
+              ]
+            },
+            {
+              exclude: [/\.mjs$/, /\.js$/, /\.html$/, /\.json$/, /\.md$/],
+              loader: require.resolve("file-loader"),
+              options: {
+                name: "static/media/[name].[hash:8].[ext]"
+              }
+            }
           ]
         }
       ]
@@ -106,25 +279,11 @@ export default async ({
     plugins: (dev
       ? []
       : [
-          new webpack.optimize.UglifyJsPlugin({
-            compress: {
-              warnings: false,
-              // Disabled because of an issue with Uglify breaking seemingly valid code:
-              // https://github.com/facebookincubator/create-react-app/issues/2376
-              // Pending further investigation:
-              // https://github.com/mishoo/UglifyJS2/issues/2011
-              comparisons: false
-            },
-            output: {
-              comments: false,
-              // Turned on because emoji and regex is not minified properly using default
-              // https://github.com/facebookincubator/create-react-app/issues/2488
-              ascii_only: true
-            },
-            sourceMap: true,
-            // Don't minify the vendor chunk, since it only contains minified modules anyway.
-            // Adds 13K to the bundle (because of webpack stuff) but speeds up the build 3x!
-            exclude: /vendor/
+          new MiniCssExtractPlugin({
+            // Options similar to the same options in webpackOptions.output
+            // both options are optional
+            filename: "static/css/[name].[contenthash:8].css",
+            chunkFilename: "static/css/[name].[contenthash:8].chunk.css"
           }),
           new ManifestPlugin({
             fileName: "asset-manifest.json"
@@ -153,9 +312,7 @@ export default async ({
       new webpack.DefinePlugin(env.stringified),
       // This is necessary to emit hot updates (currently CSS only):
 
-      ...devPlugins,
-      // Add framework-specific plugins
-      ...frameworkConfig.plugins
+      ...devPlugins
     ]),
     node: {
       fs: "empty",
@@ -166,6 +323,52 @@ export default async ({
       hints: false
     },
     optimization: {
+      minimize: !dev,
+      minimizer: [
+        // This is only used in production mode
+        new TerserPlugin({
+          terserOptions: {
+            parse: {
+              // we want terser to parse ecma 8 code. However, we don't want it
+              // to apply any minfication steps that turns valid ecma 5 code
+              // into invalid ecma 5 code. This is why the 'compress' and 'output'
+              // sections only apply transformations that are ecma 5 safe
+              // https://github.com/facebook/create-react-app/pull/4234
+              ecma: 8
+            },
+            compress: {
+              ecma: 5,
+              warnings: false,
+              // Disabled because of an issue with Uglify breaking seemingly valid code:
+              // https://github.com/facebook/create-react-app/issues/2376
+              // Pending further investigation:
+              // https://github.com/mishoo/UglifyJS2/issues/2011
+              comparisons: false,
+              // Disabled because of an issue with Terser breaking valid code:
+              // https://github.com/facebook/create-react-app/issues/5250
+              // Pending futher investigation:
+              // https://github.com/terser-js/terser/issues/120
+              inline: 2
+            },
+            mangle: {
+              safari10: true
+            },
+            output: {
+              ecma: 5,
+              comments: false,
+              // Turned on because emoji and regex is not minified properly using default
+              // https://github.com/facebook/create-react-app/issues/2488
+              ascii_only: true
+            }
+          },
+          // Use multi-process parallel running to improve the build speed
+          // Default number of concurrent runs: os.cpus().length - 1
+          parallel: true,
+          // Enable file caching
+          cache: true,
+          sourceMap: shouldUseSourceMap
+        })
+      ],
       // Automatically split vendor and commons
       // https://twitter.com/wSokra/status/969633336732905474
       // https://medium.com/webpack/webpack-4-code-splitting-chunk-graph-and-the-splitchunks-optimization-be739a861366
