+++ packages/cli/src/ambient.d.ts
@@ -1,8 +1,7 @@
-declare module "autoprefixer";
 declare module "args";
-declare module "detect-port";
 declare module "sander";
 declare module "html-webpack-plugin";
 declare module "webpack-manifest-plugin";
 declare module "react-dev-utils/*";
 declare module "friendly-errors-webpack-plugin";
+declare module "terser-webpack-plugin";
